Add shacl shape generation. Clean up code in lib/
